set global scroll bar style
